move ignore annotation to annotation dir
